Handle merge conflicts with develop
