+++ src/components/pages/dashboard/SortableVideoCard/SortableVideoCard.tsx
@@ -1,7 +1,5 @@
 import Image from "next/image";
-import { useState } from "react";
-import bookmarkFill from "/public/assets/icons/bookmark-fill.svg";
-import bookmarkIcon from "/public/assets/icons/bookmark.svg";
+import trashIcon from "/public/assets/icons/trash.svg";
 import grabIcon from "/public/assets/icons/grab.svg";
 import { useSortable } from "@dnd-kit/sortable";
 import { CSS } from "@dnd-kit/utilities";
@@ -20,7 +18,6 @@ export default function SortableVideoCard({
 	activeId,
 	handleDelete,
 }: VideoCardProps) {
-	const [isDeleted, setIsDeleted] = useState(false);
 	const { listeners, setNodeRef, setActivatorNodeRef, transform, transition } =
 		useSortable({ id: video.id, data: { video } });
 
@@ -68,26 +65,15 @@ export default function SortableVideoCard({
 					title="Delete video from playlist"
 					className={styles.card__button}
 					onClick={() => {
-						setIsDeleted(true);
 						handleDelete(video.id);
 					}}
 				>
-					{!isDeleted && (
-						<Image
-							src={bookmarkFill}
-							alt=""
-							aria-hidden="true"
-							className={styles.card__icon}
-						/>
-					)}
-					{isDeleted && (
-						<Image
-							src={bookmarkIcon}
-							alt=""
-							aria-hidden="true"
-							className={styles.card__icon}
-						/>
-					)}
+					<Image
+						src={trashIcon}
+						alt=""
+						aria-hidden="true"
+						className={styles.card__icon}
+					/>
 				</button>
 			</article>
 		</li>
