+++ src/components/pages/library/PlayListCard/PlayListCard.module.scss
@@ -1,0 +1,126 @@
+@use "@/styles/partials/variables" as *;
+@use "@/styles/partials/mixins" as *;
+
+.card {
+	padding: 1rem;
+	display: flex;
+	flex-direction: column;
+	gap: 1rem;
+
+	@include tablet {
+		padding: 0;
+	}
+
+	&__image-container {
+		display: none;
+		position: relative;
+		width: 100%;
+		aspect-ratio: 16/9;
+		margin-bottom: -0.5rem;
+		border-radius: 0.75rem;
+		background-color: rgba($color-black, 0.05);
+
+		@include tablet {
+			display: unset;
+		}
+	}
+
+	&__image {
+		border-radius: 0.75rem;
+		border: 0.5px solid #dbdbdb;
+	}
+
+	&__headerContainer {
+		display: flex;
+		justify-content: space-between;
+
+		@include tablet {
+			position: relative;
+		}
+
+		.button {
+			height: max-content;
+			border: none;
+			background-color: $color-white;
+			cursor: pointer;
+		}
+	}
+
+	&__title {
+		font-weight: $font-weight-regular;
+		font-size: 1rem;
+
+		// @include desktop {
+		// 	font-size: 1.25rem;
+		// }
+	}
+
+	&__description {
+		font-size: 0.75rem;
+		margin-top: 0.5rem;
+
+		@include tablet {
+			display: none;
+		}
+	}
+
+	&__icon {
+		width: 1rem;
+		height: 1rem;
+		vertical-align: middle;
+
+		@include tablet {
+			width: 1.5rem;
+			height: 1.5rem;
+		}
+
+		&--go {
+			@include tablet {
+				width: 1rem;
+				height: 1rem;
+			}
+		}
+	}
+
+	&__infoContainer {
+		display: flex;
+		justify-content: space-between;
+		align-items: center;
+		font-size: 0.75rem;
+
+		.info {
+			display: flex;
+			gap: 1rem;
+
+			@include tablet {
+				gap: 0.5rem;
+			}
+
+			@include desktop {
+				font-size: 0.875rem;
+			}
+		}
+
+		.button {
+			padding: 0.25rem 1rem;
+			border-radius: 2.1875rem;
+			border: 1px solid $color-black;
+			background-color: $color-white;
+			cursor: pointer;
+
+			&:hover {
+				background-color: #e2211c;
+				border-color: #e2211c;
+
+				.card__icon {
+					filter: invert(1);
+				}
+			}
+		}
+	}
+
+	&__message {
+		display: flex;
+		align-items: center;
+	}
+}
